Move get_tools_search_paths from FileSearch to Session
It only uses fields of FileSearch that are stored in Session too
